change to size_t

+++ src/split/esb/common.c
@@ -16,7 +16,7 @@ LOG_MODULE_DECLARE(zmk, CONFIG_ZMK_SPLIT_ESB_LOG_LEVEL);
 
 
 void zmk_split_esb_async_tx(struct zmk_split_esb_async_state *state) {
-    uint8_t tx_buf_len = 0;
+    size_t tx_buf_len = 0;
     uint8_t buf[CONFIG_ESB_MAX_PAYLOAD_LENGTH];
 
     if (k_sem_take(state->tx_sem, K_NO_WAIT)) {
